Issue #4314: Split and Organize Checkstyle inputs by Test for checks in design package: DesignForExtension
